Extract output and translators modules from the monolith (split PR 4/N)
- tfbs_footprinter3/output.py (new):
    target_species_hits_table_writer, sort_target_species_hits,
    top_x_greatest_hits, top_greatest_hits — post-scoring CSV writing
    and ranking for the top-X TFs used by the plotter.

- tfbs_footprinter3/translators.py (new):
    start_end_found_motif, unaligned2aligned_indexes,
    reg_position_translate, CpG, cage_position_translate,
    gtex_position_translate, gerp_positions_translate,
    gtrd_positions_translate, atac_pos_translate — genomic coordinate
    translators that convert feature locations to TSS-relative space
    for downstream scoring and plotting. All pure aside from two
    functions that read alignment files (CpG, unaligned2aligned_indexes).

The monolith continues to re-export every moved name so existing
callers (tests, hpc/, downstream users) stay unaffected.

Monolith: 2564 -> 1937 lines (-627, biggest single-PR reduction).
Also strips the now-obsolete ##-prefixed earlier versions of
gtrd_positions_translate and atac_pos_translate that still lingered.

All 21 unit tests green, ruff clean, smoke verifies 14 re-exported
names resolve via the tfbs_footprinter3.tfbs_footprinter3 path.
